requirements.txt
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
bcrypt==4.1.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
dramatiq==1.16.0
ecdsa==0.19.0
exceptiongroup==1.2.0
fastapi==0.110.1
h11==0.14.0
hiredis==2.3.2
httptools==0.6.1
idna==3.7
passlib==1.7.4
prometheus_client==0.20.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
